refactor: Change fibonacci vector data type
Change from int to double. Largest int fib before value truncation
is 1,836,311,903.

+++ pt-1-the-basics/chp-5-errors/exercises/src/first_n_fibs.cpp
@@ -35,9 +35,9 @@ void fibonacci(const int n){
   // post-conditions:
   if(n<0){error("Bad input for n");} // +ve integer check for n
 
-  std::vector<int> series = {0, 1}; // initialize with first 2 fibs - f0, f1
+  std::vector<double> series = {0, 1}; // initialize with first 2 fibs - f0, f1
   for(int i = 2; i<n; ++i){
-    int fn = series[i-1]+series[i-2];
+    double fn = series[i-1]+series[i-2];
     series.push_back(fn);
   }
   for(int i = 0; i<int(series.size()); ++i){
